Add ProfileSettingData as constants
