+++ src/context/constants/index.ts
@@ -2,3 +2,4 @@ export * from './global';
 export * from './filters';
 export * from './menuItems';
 export * from './login';
+export * from './profileSettings';
