--- src/mock/profileSettingData.ts
@@ -1,47 +1,0 @@
-import { IFilterOptionsProps } from 'context';
-
-export const EducationSettings: IFilterOptionsProps[] = [
-  {
-    label: 'Среднее',
-    value: 'Среднее',
-    isChecked: false
-  },
-  {
-    label: 'Среднее специальное',
-    value: 'Среднее специальное',
-    isChecked: false
-  },
-  {
-    label: 'Высшее',
-    value: 'Высшее',
-    isChecked: false
-  },
-];
-
-export const EnglishSettings: IFilterOptionsProps[] = [
-  {
-    label: 'Beginner/Elementary',
-    value: 'Beginner/Elementary',
-    isChecked: false
-  },
-  {
-    label: 'Pre-intermediate',
-    value: 'Pre-intermediate',
-    isChecked: false
-  },
-  {
-    label: 'Intermediate',
-    value: 'Intermediate',
-    isChecked: false
-  },
-  {
-    label: 'Upper-intermediate',
-    value: 'Upper-intermediate',
-    isChecked: false
-  },
-  {
-    label: 'Advanced',
-    value: 'Advanced',
-    isChecked: false
-  },
-];
